+++ mcp/tools_test.go
@@ -272,22 +272,6 @@ func TestRegisteredHandler_ValidatesInputBeforeClientCall(t *testing.T) {
 	assert.Contains(text, `missing required argument "id"`)
 }
 
-func TestParseServerOptions(t *testing.T) {
-	t.Run("defaults to safe tools only", func(t *testing.T) {
-		assert := assert.New(t)
-		opts, err := parseServerOptions(nil)
-		assert.NoError(err)
-		assert.False(opts.allowUnsafeTools)
-	})
-
-	t.Run("enables unsafe tools flag", func(t *testing.T) {
-		assert := assert.New(t)
-		opts, err := parseServerOptions([]string{"--allow-unsafe-tools"})
-		assert.NoError(err)
-		assert.True(opts.allowUnsafeTools)
-	})
-}
-
 func toolReq(args map[string]any) mcp.CallToolRequest {
 	return mcp.CallToolRequest{
 		Params: mcp.CallToolParams{
